tests: use run_check() instead of run() where possible
Using `run_check()` instead of a plain `run()` has the benefit that the
check for `rc != 0` is already included in this call and that `stdout`
is the only value returned (instead of a tuple of `stdout`, `stderr`
and `rc`).

This way tests will fail with an exception early, if any commands go
wrong instead of trying to continue with an undefined state.

This change tries to use `run_check()` wherever it seems fit in the
tests.

Signed-off-by: Chris Fiege <chris@tinyhost.de>